Content_prod/logics/requirements/req_init_standings.txt
beautifulsoup4==4.12.2
PyGithub==2.1.1